Item addTag and Tags
Overloaded addTag to include either Tag or String as a parameter.
Made Tags serializable again. In order to be able to save or open
Collections as files, not only does Collection have to be serializable,
but all the classes of objects that Collection contains has to be
serializable as well.